Add files via upload
The final project is providing a scenario where I am a supplier of grocery store, and I provide it with certain fruits. In addition I have to update the stores' website with the fruits available in it + sending email to the owner with the list of those fruits. This would require a python script to do it automatically. The project consists of four parts:
1- Processing images
2- Uploading images to website
3- Uploading list of fruits with their description and images to website
4- Generating PDF file of with fruit names and weights and send them through email automatically